move types to types dir
